pyproject.toml
[tool.poetry]
name = "portfolio-optimisation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
scipy = "^1.11.4"
seaborn = "^0.13.1"
numpy = "^1.26.3"
openapi = "^1.1.0"
pydantic = "^2.5.3"
openai = "^1.6.1"
aiogram = "^3.3.0"
asyncio = "^3.4.3"
ipykernel = "^6.28.0"
pandas = "^2.1.4"
textblob = "^0.17.1"
tiktoken = "^0.5.2"
yfinance = "^0.2.35"
riskfolio-lib = "^5.0.0"
benzinga = "^1.21"
ta = "^0.11.0"
plotly = "^5.19.0"
scikit-learn = "^1.4.1.post1"
inflect = "^7.0.0"
transformers = "^4.38.1"
torch = "^2.2.1"
sentence-transformers = "^2.4.0"
catboost = "^1.2.3"
lightgbm = "^4.3.0"
xgboost = "^2.0.3"
optuna = "^3.5.0"
python-dotenv = "^1.0.1"
category-encoders = "^2.6.3"
pyportfolioopt = "1.5.2"
tqdm = "^4.66.2"

[tool.poetry.group.api.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.29"
yfinance = "^0.2.37"
pandas = "^2.2.1"
beautifulsoup4 = "^4.12.3"
ta = "^0.11.0"
catboost = "^1.2.3"
scikit-learn = "^1.4.1.post1"
category-encoders = "^2.6.3"
python-dotenv = "^1.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
